The first set of changes
*Split the rotary capability into two seperate capabilities
*Changed propogation algorithm
*Changed master axis energy calculations
*Fixed some bugs
*Made it possible for gears to connect to one another by being attached
on their axis
*Added a new type of master axis (crystalline master axis). No magic
beams directly interact with gears anymore (it looked a bit weird), they
now interact with the crystalline master axis
*Changed how void will work to do the void effect even if a beam only
has a tiny amount of void.

So much code was changed that there is a good chance I created a lot of
new bugs. Bugtesting to follow, along with new features.

+++ src/main/java/com/Da_Technomancer/crossroads/API/MiscOp.java
@@ -26,7 +26,7 @@ public final class MiscOp{
 	 * 2 means goes to closest .5 value, rounding up
 	 */
 	public static double centerCeil(double numIn, int tiers){
-		return ((numIn > 0) ? Math.ceil(numIn * tiers) : Math.floor(numIn * tiers)) / tiers;
+		return ((numIn > 0) ? Math.ceil(numIn * (double) tiers) : Math.floor(numIn * (double) tiers)) / (double) tiers;
 	}
 
 	public static double findEfficiency(double speedIn, double lowerLimit, double upperLimit){
